chore: error message update
