+++ web/containers/Providers/EventHandler.tsx
@@ -92,10 +92,7 @@ export default function EventHandler({ children }: { children: ReactNode }) {
       message.content,
       message.status
     )
-    if (
-      message.status === MessageStatus.Ready ||
-      message.status === MessageStatus.Error
-    ) {
+    if (message.status !== MessageStatus.Pending) {
       // Mark the thread as not waiting for response
       updateThreadWaiting(message.thread_id, false)
