+++ web/screens/Chat/ChatBody/index.tsx
@@ -2,17 +2,9 @@ import { Fragment } from 'react'
 
 import ScrollToBottom from 'react-scroll-to-bottom'
 
-import {
-  ChatCompletionRole,
-  ConversationalExtension,
-  ExtensionType,
-  InferenceEngine,
-  MessageStatus,
-} from '@janhq/core'
+import { InferenceEngine, MessageStatus } from '@janhq/core'
 import { Button } from '@janhq/uikit'
-import { useAtomValue, useSetAtom } from 'jotai'
-
-import { RefreshCcw } from 'lucide-react'
+import { useAtomValue } from 'jotai'
 
 import LogoMark from '@/containers/Brand/Logo/Mark'
 
@@ -22,45 +14,16 @@ import { useGetDownloadedModels } from '@/hooks/useGetDownloadedModels'
 
 import { useMainViewState } from '@/hooks/useMainViewState'
 
-import useSendChatMessage from '@/hooks/useSendChatMessage'
-
 import ChatItem from '../ChatItem'
 
-import { extensionManager } from '@/extension'
-import {
-  deleteMessageAtom,
-  getCurrentChatMessagesAtom,
-} from '@/helpers/atoms/ChatMessage.atom'
-import { activeThreadAtom } from '@/helpers/atoms/Thread.atom'
+import ErrorMessage from '../ErrorMessage'
+
+import { getCurrentChatMessagesAtom } from '@/helpers/atoms/ChatMessage.atom'
 
 const ChatBody: React.FC = () => {
   const messages = useAtomValue(getCurrentChatMessagesAtom)
   const { downloadedModels } = useGetDownloadedModels()
   const { setMainViewState } = useMainViewState()
-  const thread = useAtomValue(activeThreadAtom)
-  const deleteMessage = useSetAtom(deleteMessageAtom)
-  const { resendChatMessage } = useSendChatMessage()
-
-  const regenerateMessage = async () => {
-    const lastMessageIndex = messages.length - 1
-    const message = messages[lastMessageIndex]
-    if (message.role !== ChatCompletionRole.User) {
-      // Delete last response before regenerating
-      deleteMessage(message.id ?? '')
-      if (thread) {
-        await extensionManager
-          .get<ConversationalExtension>(ExtensionType.Conversational)
-          ?.writeMessages(
-            thread.id,
-            messages.filter((msg) => msg.id !== message.id)
-          )
-      }
-      const targetMessage = messages[lastMessageIndex - 1]
-      if (targetMessage) resendChatMessage(targetMessage)
-    } else {
-      resendChatMessage(message)
-    }
-  }
 
   if (downloadedModels.length === 0)
     return (
@@ -118,26 +81,10 @@ const ChatBody: React.FC = () => {
           {messages.map((message, index) => (
             <div key={message.id}>
               <ChatItem {...message} key={message.id} />
-              {message.status === MessageStatus.Error &&
+              {(message.status === MessageStatus.Error ||
+                message.status === MessageStatus.Stopped) &&
                 index === messages.length - 1 && (
-                  <div
-                    key={message.id}
-                    className="mt-10 flex flex-col items-center"
-                  >
-                    <span className="mb-3 text-center text-sm font-medium text-gray-500">
-                      Oops! The generation was interrupted. Let&apos;s give it
-                      another go!
-                    </span>
-                    <Button
-                      className="w-min"
-                      themes="outline"
-                      onClick={regenerateMessage}
-                    >
-                      <RefreshCcw size={14} className="" />
-                      <span className="w-2" />
-                      Regenerate
-                    </Button>
-                  </div>
+                  <ErrorMessage message={message} />
                 )}
             </div>
           ))}
