+++ web/screens/Chat/SimpleTextMessage/index.tsx
@@ -104,10 +104,7 @@ const SimpleTextMessage: React.FC<ThreadMessage> = (props) => {
   }, [])
 
   useEffect(() => {
-    if (
-      props.status === MessageStatus.Ready ||
-      props.status === MessageStatus.Error
-    ) {
+    if (props.status !== MessageStatus.Pending) {
       return
     }
     const currentTimestamp = new Date().getTime() // Get current time in milliseconds
